backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.14
Pillow==10.1.0
python-magic==0.4.27
aiofiles==23.2.1
pydantic==2.5.2
typing-extensions==4.8.0